requirements.txt
autopep8==1.5.6
bcrypt==3.2.0
certifi==2020.12.5
cffi==1.14.5
click==7.1.2
cloudinary==1.25.0
dnspython==2.1.0
email-validator==1.1.2
fastapi==0.63.0
gunicorn==20.1.0
h11==0.12.0
idna==3.1
numpy==1.20.3
opencv-python-headless==4.5.2.52
passlib==1.7.4
Pillow==8.2.0
pycodestyle==2.7.0
pycparser==2.20
pydantic==1.8.1
PyJWT==2.0.1
pymongo==3.11.3
python-dotenv==0.15.0
python-multipart==0.0.5
six==1.15.0
starlette==0.13.6
tflite-runtime==2.5.0
toml==0.10.2
typing-extensions==3.7.4.3
urllib3==1.26.4
uvicorn==0.13.4
